+++ R/ternary_box.R
@@ -10,7 +10,7 @@ ternary_box <- function(lty = "solid", ...) {
   col <- list(...)$col %||% graphics::par("fg")
   lwd <- list(...)$lwd %||% 1
 
-  graphics::polygon(x = c(0, 0.5, 1), y = c(0, .top, 0),
+  graphics::polygon(x = c(0, 0.5, 1), y = c(0, the$top, 0),
                     border = col, lty = lty, lwd = lwd)
 
   invisible(NULL)