+++ R/zzz.R
@@ -1,9 +1,7 @@
 .onLoad <- function(libname, pkgname) {
   op <- options()
   op.isopleuros <- list(
-    isopleuros.missing = FALSE,
-    isopleuros.center = NULL,
-    isopleuros.scale = NULL
+    isopleuros.missing = FALSE
   )
   toset <- !(names(op.isopleuros) %in% names(op))
   if(any(toset)) options(op.isopleuros[toset])